requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==4.2
django-cors-headers==4.4.0
django-jazzmin==3.0.0
django-modeltranslation==0.19.7
djangorestframework==3.15.2
drf-yasg==1.21.7
idna==3.7
inflection==0.5.1
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
setuptools==72.2.0
sqlparse==0.5.1
uritemplate==4.1.1
urllib3==2.2.2
